Modify chatting module structure
