+++ app/plain/route.js
@@ -33,7 +33,6 @@ router.use(Manager.route);     // to view log module
 // it uses common feature for each admin and manager, then assign in app router.
 // other features use each module's router. eg, modifying account records or log records
 router.get(routeTable.account_root + routeTable.account.signOut, Account.signOut);
-router.post(routeTable.account_root + routeTable.account.signIn, Account.signIn);
 router.post(routeTable.account_root + routeTable.account.registerSimple, Account.registerSimple);
 
 // need to place down here for except admin page log
