+++ core/config/route_default.json
@@ -40,10 +40,14 @@
     "message": "/register",
     "reply": "/register/reply"
   },
+  "chatting_root": "/chatting",
   "teamblog_root": "/blog",
   "teamblog": {
     "list": "/",
     "write": "/write"
   },
-  "teamblog_post": "/post"
+  "teamblog_post": "/post",
+  "teamblog_tag": {
+    "list": "/tag"
+  }
 }
