+++ module/chatting/route.js
@@ -7,6 +7,6 @@ var chatting = require('./lib/chatting');
 var router = express.Router();
 var routeTable = misc.getRouteTable();
 
-router.get(routeTable.chatting.root, chatting.index);
+router.get(routeTable.chatting_root, chatting.index);
 
 module.exports = router;
