+++ theme/plain/page/chatting.html
@@ -71,7 +71,7 @@
         $('#messages').append($('<li>').text(data.nickname + ' : ' + data.msg));
     });
 
-    console.log(user);
+    console.log(global);
 </script>
 
 {% include '../include/footer.html' %}
